+++ RELEASE_NOTES.md
@@ -11,3 +11,7 @@
 * Improved project configuration
 * Fixed a bug where the disassembler would obtain instruction details when they were disabled
 * Changed `X86` and `PowerPC` to `X86Syntax` and `PowerPCSyntax`
+
+### 0.0.4 - July 24 2017
+* Removed `detailsOn` constructor parameter; use the `Details` property instead
+* Added x86-specific unit tests